requirements.txt
Flask>=2.2.5
gunicorn>=20.1.0
# Postgres driver for deploying to Neon/Postgres
psycopg2-binary>=2.9
# Optional: load .env files during local development
python-dotenv>=1.0.0
